Create prims.c
prims doesn't work for directed graph perfectly. it is used to find minimum spanning tree. if there are more directefd edges there is a chance to form cycle which is not allowed. but here we get bit perfect solution.
Enter the number of nodes and edges: 5
9
Enter the edges and their weights: 0 1 4
0 2 2
1 3 2
1 2 3
1 2 -1
1 4 3
2 4 5
4 3 -5
2 3 4
Enter the starting node: 0
Edge 	Weight
0 - 1 	4
0 - 2 	2
1 - 3 	2
1 - 4 	3