Create kruskals.py
3 - 2: -5
4 - 2: 1
0 - 4: 2
1 - 2: 2
here we take the smallest distance first so -5 is selected first. kruskals also does not work for directed graph so it gives a WRONG ANSWER.

i used python here because it is easy. i learned dsa in python in this holiday.
this gives weight as 0.
prims and kruskals does not work for directed graph