Create dihkstra.py
dijkstra does not work for -ve weights. it is better to use bellman ford's algorithm instead to get perfect optimal solution.
i used python here because it is easy to use. i know data structures in c and python only. i used python here because it consumes less time to type and i am bit slow in typing. to manage time i did it in python